fs/ntfs3: Fix compile-time warnings in run_unpack()
../fs/ntfs3/run.c:940:8: error: comparison of distinct pointer types ('typeof (prev_lcn) *' (aka 'unsigned long long *') and 'typeof (dlcn) *' (aka 'long long *')) [-Werror,-Wcompare-distinct-pointer-types]
                        if (check_add_overflow(prev_lcn, dlcn, &lcn))
                            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../include/linux/overflow.h:70:15: note: expanded from macro 'check_add_overflow'
        (void) (&__a == &__b);                  \
                ~~~~ ^  ~~~~

+++ fs/ntfs3/run.c
@@ -940,7 +940,7 @@ int run_unpack(struct runs_tree *run, struct ntfs_sb_info *sbi, CLST ino,
 			/* Check special combination: 0 + SPARSE_LCN64. */
 			if (!prev_lcn && dlcn == SPARSE_LCN64) {
 				lcn = SPARSE_LCN64;
-			} else if (check_add_overflow(prev_lcn, dlcn, &lcn)) {
+			} else if (check_add_overflow(prev_lcn, (u64)dlcn, &lcn)) {
 				return -EINVAL;
 			}
 			prev_lcn = lcn;
